Templete alterations and this component was moved to shared to be used in a generic way
