--- angular/pos/src/app/components/user/modal-delete/modal-delete.component.css
@@ -1,23 +1,0 @@
-.container{
-    text-align: center;
-}
-
-.title{
-    font-size: 24px;
-}
-.btn{
-    padding-left: 30px;
-    padding-right: 30px;
-}
-
-#btn-confirm{
-  background-color: #3f51b5;
-  border: none;
-  border: 3px solid #3f51b5;
-}
-
-#btn-back{
-  background-color: #787878;
-  border: none;
-  border: 3px solid #787878;
-}
