--- angular/pos/src/app/components/user/modal-delete/modal-delete.component.html
@@ -1,28 +1,0 @@
-<div class="container">
-
-  <div class="container">
-
-    <h1>Confirmar</h1>
-    <p>Você deseja realmente deletar?</p>
-
-  </div>
-
-  <div class="row align-items-center">
-
-    <div class="col">
-      <button type="button" id="btn-confirm" class="btn btn-lg btn-primary ml-4" (click) = "delete()"> Sim </button>
-    </div>
-    <div class="col">
-      <button type="button" id="btn-back" class="btn btn-lg btn-secondary ml-4" (click) = "goToList()"> Não </button>
-
-    </div>
-
-
-  </div>
-
-</div>
-
-
-
-
-
