+++ angular/pos/src/app/components/user/user-list/user-list.component.ts
@@ -1,4 +1,4 @@
-import { ModalDeleteComponent } from './../modal-delete/modal-delete.component';
+
 import { AppConstants } from './../../../app-constants';
 import { MatDialog } from '@angular/material/dialog';
 import { Component, OnInit } from '@angular/core';
@@ -6,6 +6,7 @@ import { Router } from '@angular/router';
 import { UserService } from '../user.service';
 import { UserEditComponent } from '../user-edit/user-edit.component';
 import {UserViewModalComponent} from '../user-view-modal/user-view-modal.component';
+import {ModalDeleteComponent} from '../../../shared/modal-delete/modal-delete.component';
 
 @Component({
   selector: 'app-user-list',
