+++ angular/pos/src/app/shared/modal-delete/modal-delete.component.ts
@@ -1,4 +1,4 @@
-import { UserService } from './../user.service';
+import { UserService } from '../../components/user/user.service';
 import { Component, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
